requirements.txt
Flask==3.1.0
pymongo==4.10.1
gunicorn==23.0.0
python-dotenv
certifi